test(generation_tasks): 拆分 4042 行 test_generation_tasks_service.py 为按被测对象分文件
按 execute_video_task/execute_generation_task/formal image finalization 等
机械切分，跨文件复用的 _FakePM/_prepare_files/_ad_pm 等基础设施迁入本目录
conftest.py。不改用例语义；拆分前后 collect-only 清单以 nodeid 后缀多重集
校验一致（125 项）。
